Add gyro control functions and revised
in order to get up the ramp, i have to use the gyro to correct
The gyro drifts every second and i needed to subtract .135 for every
second to get an accurate gyro and compensate for the drift
The goForwardGyro and the goBackwardGyro are unique because they use
basically exact gyro readings. I slowed one of the other sides down in
the while loop
The backward uses a different compensation than forward so it has less
trouble getting up the ramp.
It has trouble getting up the ramp because the robot slips on the ramp.
I created a ramp function that goes up the ramp and refined other
functions like goDump.